added fixes

+++ actions/actions.js
@@ -1,5 +1,7 @@
 "use server";
+import { authOptions } from "@/app/api/auth/[...nextauth]/route";
 import { PrismaClient } from "@prisma/client";
+import { getServerSession } from "next-auth";
 import { revalidatePath } from "next/cache";
 import { redirect } from "next/navigation";
 const prisma = new PrismaClient();
@@ -24,34 +26,38 @@ export const fetchSingleBlog = async (id) => {
 
 export const addBlog = async (formData) => {
 
-
     // collect info from form using formData
     const imageUrl = formData.get('imageUrl');
     const title = formData.get('title');
     const category = formData.get('category');
     const description = formData.get('description');
 
+    // session to get the current logged in user info
 
-    // push the data into the DB
-    const new_blog = await prisma.blog.create({
-        data: {
-            imageUrl: imageUrl ? imageUrl : null,
-            title,
-            category,
-            description
-        }
-    })
+    const session = await getServerSession(authOptions);
 
-    revalidatePath('/blogs/add-blog')
-    redirect('/blogs')
+    // only admin can add blog
+    if (session?.user?.role === 'ADMIN' || session?.user?.permissions?.includes('CREATE_BLOG')) {
+
+        // push the data into the DB
+        const new_blog = await prisma.blog.create({
+            data: {
+                imageUrl: imageUrl ? imageUrl : null,
+                title,
+                category,
+                description,
+                authorId: session?.user?.id
+            }
+        })
+
+        revalidatePath('/blogs/add-blog')
+        redirect('/blogs')
+    }
 
 
 }
 
-
-
 // Update a Blog
-
 
 export const updateBlog = async (id, formData) => {
 
@@ -61,22 +67,35 @@ export const updateBlog = async (id, formData) => {
     const category = formData.get('category');
     const description = formData.get('description');
 
+    // session 
 
-    // push the data into the DB
-    const updated_blog = await prisma.blog.update({
-        where: {
-            id: id,
-        },
-        data: {
-            imageUrl: imageUrl ? imageUrl : null,
-            title,
-            category,
-            description
-        }
-    })
+    const session = await getServerSession(authOptions);
 
-    revalidatePath(`/blogs/update-blog/${id}`)
-    redirect('/blogs')
+    // only admin can add blog
+    if (session?.user?.role === 'ADMIN' || session?.user?.permissions?.includes('EDIT_BLOG')) {
+        // push the data into the DB
+        const updated_blog = await prisma.blog.update({
+            where: {
+                id: id,
+            },
+            data: {
+                imageUrl: imageUrl ? imageUrl : null,
+                title,
+                category,
+                description,
+                authorId: session?.user?.id
+            }
+        })
+
+        revalidatePath(`/blogs/update-blog/${id}`)
+        redirect('/blogs')
+
+    } else {
+        console.log('Not Possible!')
+    }
+
+
+
 }
 
 
@@ -84,9 +103,15 @@ export const updateBlog = async (id, formData) => {
 export const addCommentToBlog = async (blogId, formData) => {
     // collect info from form using formData
     const text = formData.get('text');
+
+    // session
+
+    const session = await getServerSession(authOptions);
+
     // push the data into the DB
     const added_comment = await prisma.comment.create({
         data: {
+            authorId: session?.user?.id,
             blogId: blogId,
             text: text
         }
@@ -122,12 +147,80 @@ export const fetchComments = async (blogId) => {
 
 export const deleteComment = async (commentId, blogId) => {
 
-    await prisma.comment.delete({
+    // only auther of this comment can delete it!
+
+    const session = await getServerSession(authOptions);
+
+    // comment of autherId
+
+    const commentData = await prisma.comment.findFirst({
         where: {
             id: commentId
         }
     })
 
-    revalidatePath(`/blogs/${blogId}`)
+
+    if (session?.user?.id === commentData?.authorId) {
+        await prisma.comment.delete({
+            where: {
+                id: commentId
+            }
+        })
+
+        revalidatePath(`/blogs/${blogId}`)
+    } else {
+        console.log('You Are Not Authorize to Delete This Comment!');
+    }
 
 }
+
+
+// fetch users
+
+export const fetchUsers = async () => {
+    const users = await prisma.user.findMany({
+        take: 5
+    })
+
+    return users
+
+}
+
+// assign user to a particular role from admin panel
+
+export const assignPermission = async (userId, formData) => {
+
+    const permission_name = formData.get('permission_name');
+
+    // session 
+    const session = await getServerSession(authOptions);
+
+    // only admin can add blog
+    if (session?.user?.role === 'ADMIN') {
+
+        // asssign some permission to user by admin
+        const assigned_user = await prisma.user.update({
+            where: {
+                id: userId,
+            },
+
+            data: {
+                permissions: {
+                    push: permission_name
+                }
+
+            }
+
+        })
+        revalidatePath(`/admin/dashboard`)
+        redirect(`/admin/dashboard`)
+
+    }
+
+
+
+}
+
+
+
+
